mcp-client-langchain/pyproject.toml
[project]
name = "mcp-client-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.4",
    "langgraph>=0.4.7",
    "requests>=2.32.3",
]
